handle files and apply psr

+++ app/Http/Resources/OrderItemResource.php
@@ -9,17 +9,17 @@ class OrderItemResource extends JsonResource
     public function toArray($request)
     {
         return [
-            "id"           => $this->id,
-            "coupon_num"   => (string)$this->coupon_num,
-            "amount"       => (double)$this->amount,
-            "quantity"     => (int)$this->quantity,
-            "order_id"     => $this->order_id,
-            "appointment"  => $this->appointment,
-            "qr_code_link" => $this->qr_code_link,
+            'id'           => $this->id,
+            'coupon_num'   => (string) $this->coupon_num,
+            'amount'       => (float) $this->amount,
+            'quantity'     => (int) $this->quantity,
+            'order_id'     => $this->order_id,
+            'appointment'  => $this->appointment,
+            'qr_code_link' => $this->qr_code_link,
 
-            "product"      => ProductResource::make($this->whenLoaded('product')),
+            'product'      => ProductResource::make($this->whenLoaded('product')),
 //            "offer"        => OfferResource::make($this->whenLoaded('offer')),
-            "vendor"       => VendorResource::make($this->whenLoaded('vendor')),
+            'vendor'       => VendorResource::make($this->whenLoaded('vendor')),
         ];
     }
 }
