+++ app/Http/Resources/OrderResource.php
@@ -11,16 +11,16 @@ class OrderResource extends JsonResource
     {
         return [
             'id'          => $this->id,
-            'total'       => (double)$this->total,
+            'total'       => (float) $this->total,
             'status'      => $this->status,
-            'note'        => (string)$this->note,
-            'name'        => (string)$this->name,
+            'note'        => (string) $this->note,
+            'name'        => (string) $this->name,
 
-            'phone'       => (string)$this->phone,
-            'is_used'     => (bool)$this->is_used,
-            'invoice_url' => (string)$this->invoice_url,
+            'phone'       => (string) $this->phone,
+            'is_used'     => (bool) $this->is_used,
+            'invoice_url' => (string) $this->invoice_url,
 
-            'cancellation_reason' => (string)$this->cancellation_reason,
+            'cancellation_reason' => (string) $this->cancellation_reason,
 
             'created_at'  => Carbon::parse($this->created_at)->diffForHumans(),
             'updated_at'  => Carbon::parse($this->updated_at)->diffForHumans(),
